--- src/test/java/codesquad/fineants/spring/api/portfolio_stock/PortfolioHoldingRestControllerTest.java
@@ -1,194 +1,0 @@
-package codesquad.fineants.spring.api.portfolio_stock;
-
-import static org.hamcrest.Matchers.*;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.BDDMockito.anyLong;
-import static org.mockito.BDDMockito.*;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.*;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.MediaType;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-
-import codesquad.fineants.domain.member.Member;
-import codesquad.fineants.domain.oauth.support.AuthMember;
-import codesquad.fineants.domain.oauth.support.AuthPrincipalArgumentResolver;
-import codesquad.fineants.domain.portfolio.Portfolio;
-import codesquad.fineants.domain.portfolio.PortfolioRepository;
-import codesquad.fineants.domain.portfolio_holding.PortfolioHolding;
-import codesquad.fineants.domain.stock.Market;
-import codesquad.fineants.domain.stock.Stock;
-import codesquad.fineants.spring.api.errors.handler.GlobalExceptionHandler;
-import codesquad.fineants.spring.api.portfolio_stock.controller.PortfolioStockRestController;
-import codesquad.fineants.spring.api.portfolio_stock.manager.SseEmitterManager;
-import codesquad.fineants.spring.api.portfolio_stock.request.PortfolioStockCreateRequest;
-import codesquad.fineants.spring.api.portfolio_stock.response.PortfolioStockCreateResponse;
-import codesquad.fineants.spring.api.portfolio_stock.service.PortfolioStockService;
-import codesquad.fineants.spring.auth.HasPortfolioAuthorizationAspect;
-import codesquad.fineants.spring.config.JpaAuditingConfiguration;
-import codesquad.fineants.spring.config.SpringConfig;
-
-@ActiveProfiles("test")
-@WebMvcTest(controllers = PortfolioStockRestController.class)
-@Import(value = {SpringConfig.class, HasPortfolioAuthorizationAspect.class})
-@MockBean(JpaAuditingConfiguration.class)
-class PortfolioHoldingRestControllerTest {
-
-	private MockMvc mockMvc;
-
-	@Autowired
-	private PortfolioStockRestController portfolioStockRestController;
-
-	@Autowired
-	private GlobalExceptionHandler globalExceptionHandler;
-
-	@Autowired
-	private ObjectMapper objectMapper;
-
-	@MockBean
-	private AuthPrincipalArgumentResolver authPrincipalArgumentResolver;
-
-	@MockBean
-	private PortfolioStockService portfolioStockService;
-
-	@MockBean
-	private PortfolioRepository portfolioRepository;
-
-	@MockBean
-	private SseEmitterManager manager;
-
-	private Member member;
-	private Portfolio portfolio;
-	private Stock stock;
-
-	private PortfolioHolding portfolioHolding;
-
-	@BeforeEach
-	void setup() {
-		mockMvc = MockMvcBuilders.standaloneSetup(portfolioStockRestController)
-			.setControllerAdvice(globalExceptionHandler)
-			.setCustomArgumentResolvers(authPrincipalArgumentResolver)
-			.alwaysDo(print())
-			.build();
-
-		given(authPrincipalArgumentResolver.supportsParameter(any())).willReturn(true);
-
-		member = Member.builder()
-			.id(1L)
-			.nickname("일개미1234")
-			.email("kim1234@gmail.com")
-			.provider("local")
-			.password("kim1234@")
-			.profileUrl("profileValue")
-			.build();
-
-		AuthMember authMember = AuthMember.from(member);
-
-		given(authPrincipalArgumentResolver.resolveArgument(any(), any(), any(), any())).willReturn(authMember);
-
-		portfolio = Portfolio.builder()
-			.id(1L)
-			.name("내꿈은 워렌버핏")
-			.securitiesFirm("토스")
-			.budget(1000000L)
-			.targetGain(1500000L)
-			.maximumLoss(900000L)
-			.member(member)
-			.build();
-
-		stock = Stock.builder()
-			.tickerSymbol("005930")
-			.companyName("삼성전자보통주")
-			.companyNameEng("SamsungElectronics")
-			.stockCode("KR7005930003")
-			.market(Market.KOSPI)
-			.build();
-
-		portfolioHolding = PortfolioHolding.builder()
-			.id(1L)
-			.currentPrice(null)
-			.portfolio(portfolio)
-			.stock(stock)
-			.build();
-	}
-
-	@DisplayName("사용자는 포트폴리오에 종목을 추가한다")
-	@Test
-	void addPortfolioStock() throws Exception {
-		PortfolioStockCreateResponse response = PortfolioStockCreateResponse.from(
-			PortfolioHolding.empty(portfolio, stock));
-		given(portfolioRepository.findById(anyLong())).willReturn(Optional.of(portfolio));
-		given(portfolioStockService.addPortfolioStock(anyLong(), any(PortfolioStockCreateRequest.class),
-			any(AuthMember.class))).willReturn(response);
-
-		Map<String, Object> requestBodyMap = new HashMap<>();
-		requestBodyMap.put("tickerSymbol", stock.getTickerSymbol());
-
-		String body = objectMapper.writeValueAsString(requestBodyMap);
-		Long portfolioId = portfolio.getId();
-		// when & then
-		mockMvc.perform(post("/api/portfolio/" + portfolioId + "/holdings")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(body))
-			.andExpect(status().isCreated())
-			.andExpect(jsonPath("code").value(equalTo(201)))
-			.andExpect(jsonPath("status").value(equalTo("Created")))
-			.andExpect(jsonPath("message").value(equalTo("포트폴리오 종목이 추가되었습니다")))
-			.andExpect(jsonPath("data").value(equalTo(null)));
-	}
-
-	@DisplayName("사용자는 포트폴리오에 종목을 추가할때 stockId를 필수로 같이 전송해야 한다")
-	@Test
-	void addPortfolioStockWithStockIdIsNull() throws Exception {
-		Map<String, Object> requestBodyMap = new HashMap<>();
-		requestBodyMap.put("stockId", null);
-
-		String body = objectMapper.writeValueAsString(requestBodyMap);
-		Long portfolioId = portfolio.getId();
-
-		given(portfolioRepository.findById(anyLong())).willReturn(Optional.of(portfolio));
-
-		// when & then
-		mockMvc.perform(post("/api/portfolio/" + portfolioId + "/holdings")
-				.contentType(MediaType.APPLICATION_JSON)
-				.content(body))
-			.andExpect(status().isBadRequest())
-			.andExpect(jsonPath("code").value(equalTo(400)))
-			.andExpect(jsonPath("status").value(equalTo("Bad Request")))
-			.andExpect(jsonPath("message").value(equalTo("잘못된 입력형식입니다")))
-			.andExpect(jsonPath("data").isArray());
-	}
-
-	@DisplayName("사용자는 포트폴리오 종목을 삭제한다")
-	@Test
-	void deletePortfolioStock() throws Exception {
-		// given
-		Long portfolioHoldingId = portfolioHolding.getId();
-		Long portfolioId = portfolio.getId();
-
-		given(portfolioRepository.findById(anyLong())).willReturn(Optional.of(portfolio));
-		// when & then
-		mockMvc.perform(delete("/api/portfolio/" + portfolioId + "/holdings/" + portfolioHoldingId))
-			.andExpect(status().isOk())
-			.andExpect(jsonPath("code").value(equalTo(200)))
-			.andExpect(jsonPath("status").value(equalTo("OK")))
-			.andExpect(jsonPath("message").value(equalTo("포트폴리오 종목이 삭제되었습니다")))
-			.andExpect(jsonPath("data").value(equalTo(null)));
-	}
-}
